+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/StatusController.java
@@ -1,5 +1,7 @@
 package gov.healthit.chpl.web.controller;
 
+import org.apache.log4j.LogManager;
+import org.apache.log4j.Logger;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -11,12 +13,14 @@ import io.swagger.annotations.ApiOperation;
 @Api
 @RestController
 public class StatusController {
-	
+	private static final Logger logger = LogManager.getLogger(StatusController.class);
+
 	@ApiOperation(value="Check that the rest services are up and running.", 
 			notes="")
 	@RequestMapping(value="/status", method=RequestMethod.GET,
 			produces="application/json; charset=utf-8")
 	public @ResponseBody String getStatus() {
+		logger.warn("/status called");
 		return "{\"status\": \"OK\"}";
 	}
 }